requirement.txt
python-dotenv
langchain
openai
pypdf
tiktotoken
llama-cpp-python